Added a Twig_Node_Extra.
This node can be used to add some methods, with the use of userland
Node and userland Node Visitor, to the final template.

This can be usefull, for exemple, if you want to collect some datas
during the compilation with a node visitor and you want to expose
them in the final template with custom methods.