Added template information on debug mode
On debug mode (with `template_info` set to `true`) all included templates are wrapped by the path information for the template.
This is useful for easily moving in a code base which is not very well know to the developer (ie, a legacy code base or a new developer joining a project).

Example output:

    <!-- AcmeDemoBundle:Footer:Default.html.twig -->
    <div class="footer container">
    ...
    </div>

    <!-- // AcmeDemoBundle:Footer:Default.html.twig -->

The default format is (for html):

    <!-- %1$s -->%3$s%2$s%3$s<!-- // %1$s -->%3$s

Given to sprintf with the following arguments:

    $templatePath, $contents, PHP_EOL

Being `$templatePath` the template path (not resolved to file system, but in the format ie, `AcmeDemoBundle:Header:section/template.html.twig`).

`$contents` is the contents itself of the template already handled by doDisplay (ie, rendered).

`PHP_EOL` is only for formatting.

The format is determined by the option 'autoescape' passed in `$options` to `Twig_Environment`; and can be either a string or a callback.
Available format (by default) are 'html', 'js' and 'css'; adding, removing and changing formats is possible through the `Twig_Extension_Debug` extension.

The configuration argument `debug` and `template_info` must be both in true to enable this functionallity.