Clean output buffers when an exception occured
When I was testing rendering of a specific block (`renderBlock()`) of a faulty Twig template (non-existent variable) in PHPUnit, it stated that particular test was risky.

After some debugging, I found out that when an exception occurs in the `displayBlock()` method, the output buffer that was opened in `renderBlock()` would still be open. PHPUnit complains about that by saying the test was risky.

I fixed this by closing all open output buffers when an exception occurs. The fix is the same as the one that is already in place in `render()`. Next, I also applied it to `renderParentBlock()`, that has the same behaviour as `renderBlock()`.