Add gitattributes w/ export-ignore
This will allow for a `production` ready version of Twig and while still allowing a development version within the repository.
The `export-ignore` excludes un-needed files and directories when a branch is archived for release and are untuched when the
repo is cloned/forked.

This would be very useful because `releases` are intended for production use and when people are using a package (via composer
or zip), they most likely are not interested in downloading the source code as a whole into production. Thus saving disk space
and meeting security protocol.

  - e.g. /test folder, or your .travis.yml file, etc.