Clarify replace documentation and example
At a quick glance the description and example for the replace filter make it seem like the filter only replaces using a special %foo% syntax rather than working on any arbitrary search and replace. See some of the confusion here:

https://stackoverflow.com/questions/4943880/str-replace-in-twig

I tweaked the description of the filter to more closely match the PHP documentation's description of str_replace, and dropped the % formatting from the example.